Several changes after work :D 1. Switch amount to int from float and convert to milliunits that's expected by YNAB 2. Update category dne error message to suggest checking the case 3. Update gross pay offset category to be something other than the income group (because that doesn't work) 4. Grab the check date and use that for the transaction date instead of today()

+++ payslip2budget/models/transaction_base.py
@@ -8,4 +8,4 @@ class Transaction:
     payee: str
     category: Optional[str] = None
     memo: str
-    amount: float  # In dollars for base class
+    amount: int  # In dollars for base class
